pyproject.toml
[tool.poetry]
name = "tic-tac-toe"
version = "0.1.0"
description = "Tic-Tac-Toe game"
authors = ["Sergey Ilin <ilin_sergey90@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pygame = "^2.5.2"
pytest-cov = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
